Fix datatypes in PDB parsing

+++ openfold/data/data_pipeline.py
@@ -147,8 +147,8 @@ def make_protein_features(
     all_atom_positions = protein_object.atom_positions
     all_atom_mask = protein_object.atom_mask
 
-    pdb_feats["all_atom_positions"] = all_atom_positions
-    pdb_feats["all_atom_mask"] = all_atom_mask
+    pdb_feats["all_atom_positions"] = all_atom_positions.astype(np.float32)
+    pdb_feats["all_atom_mask"] = all_atom_mask.astype(np.float32)
 
     pdb_feats["resolution"] = np.array([0.]).astype(np.float32)
     pdb_feats["is_distillation"] = np.array(
